feat: tie locations to a registryId
This allows admins to set oclc RegistryId for a
specific locations (library branches). The
RegistryId do not have to be unique so that if
some locations share the same, they may be set as
such.

These registryIds are used whenever a query is
sent from Aspen Discovery to the Resource Sharing
API and serve to identify which location the query
was sent from.