DIS-239: refactor: use Evergreen username duplicate check
Aspen Discovery does not have access to all
usernames stored in Evergreen at the moment
when a patron attempts to update their usernames.

This commit removes the username duplicate check
method in favour of relying on Evergreen's API
responses, while retaining the ability to inform
the patron if the new username could not be saved
due to it being a duplicate.

Note: the username duplicate check in Evergreen
is case-sensitive.