DIS-1345: feat: display pre hold submission fee message
Currently, this feature is only available for the
Koha ILS.
Note that while there is no setting to toggle this
in Aspen, hold fees are normally displayed to Koha
patrons on the Koha OPAC if they do exist.
This is getting Aspen to match existing Koha OPAC
behaviour, depending on the library's specific
Koha configuration.